removed __init__.py files from .gitignore, added try/except blocks for reading/writing files in json_handler.py

+++ src/json_handler.py
@@ -1,4 +1,4 @@
-import sys
+import os
 import json
 
 from data.card_data import CARD_DATA
@@ -8,13 +8,25 @@ key_list = ["index", "question", "answer"]
 card_dict = { str(c[0]): dict(zip(key_list, c)) for c in CARD_DATA }
 
 
-def writes():
-    with open('data/card_data.json', "w", encoding="utf-8") as cards:
-        json.dump(card_dict, cards, ensure_ascii=False, indent=4)
-    return
+def write_file(path='data/card_data.json'):
+    if os.path.exists(path):
+        try:
+            with open(path, "w", encoding="utf-8") as cards:
+                json.dump(card_dict, cards, ensure_ascii=False, indent=4)
+            return "save successful"
+        except:
+            return "Error: open() or json.dump() failed"
 
-def reads():
-    with open('data/card_data.json', encoding="utf-8") as cards:
-        data = json.load(cards)
-    return data
+def read_file(path='data/card_data.json'):
+    if os.path.exists(path):
+        try:
+            with open(path, encoding="utf-8") as cards:
+                data = json.load(cards)
+            return data
+        except Exception as e:
+            return f"Error: {e}"
+
+
+if __name__ == "__main__":
+    writes()
